add warning for gyp failures
When upgrading minor versions packages using node-gyp often fail to
compile initially.  Until the comunity responds and packages are
updated, node-v0.12 will be unusable to some.  Most will miss
this notice, but hopefully it'll save a few headaches.

This is my minimum expectation of warning.  Ideally it would be
placed directly under "Current version: v0.12.0", and have a
light yellow background.

This warning is intended to be temporary, and be removed in a
few months.  It should then reappear when this page is upgraded
for node-v0.14.