footer: link license on master
The current website links to the license of the version branch
of node which was used to create the website.

Currently we have at least two different links deployed in
production: in the footer of `https://nodejs.org/community/` the
link to the licene points to
`https://raw.github.com/joyent/node/v0.10.36/LICENSE` and the link
on `https://nodejs.org/about/core-team/meetings/` points to
`https://raw.github.com/joyent/node/v0.10.37/LICENSE`.

This leads to the following problem:

it is quite hard to find out what is deployed in production vs.
what changes are on my local branch. Additionally, not the latest
version of the license file is linked and we even have different
versions linked in production. For the website the link should
always point to master (but for version specific api docs, to the
version that the api docs show)